Create utils file and add expense endpoints to server.py
