+++ src/api/expenses.py
@@ -3,43 +3,9 @@ from fastapi import APIRouter, HTTPException
 from pydantic import BaseModel
 
 from src import database as db
+from src import sql_utils as utils
 
 router = APIRouter()
-
-# We can probably move these to separate files?
-def get_user(user_id: int):
-    with db.engine.connect() as conn:
-        user = conn.execute(
-            sqlalchemy.text("SELECT * FROM user WHERE user_id = :user_id"),
-            [{"user_id": user_id}]
-        ).fetchone()
-        if user is None:
-            raise HTTPException(status_code=404, detail="user not found.")
-    return user
-
-
-def get_expense(user_id: int, expense_id: int):
-    with db.engine.connect() as conn:
-        expense = conn.execute(
-            sqlalchemy.text(
-                "SELECT * FROM expense WHERE user_id = :user_id AND expense_id = :expense_id"),
-            [{"user_id": user_id, "expense_id": expense_id}]
-        ).fetchone()
-        if expense is None:
-            raise HTTPException(status_code=404, detail="expense not found.")
-    return expense
-
-
-def get_category(user_id: int, category_id: int):
-    with db.engine.connect() as conn:
-        category = conn.execute(
-            sqlalchemy.text(
-                "SELECT * FROM category WHERE user_id = :user_id AND category_id = :category_id"),
-            [{"user_id": user_id, "category_id": category_id}]
-        ).fetchone()
-        if category is None:
-            raise HTTPException(status_code=404, detail="category not found.")
-    return category
 
 
 @router.get("/user/{user_id}/expense/{expense_id}", tags=["expenses"])
@@ -54,10 +20,10 @@ def get_expense(user_id: int, expense_id: int):
     - `category`: the user defined category of the item
     - `description`: the user defined description of the item
     """
-    user = get_user(user_id)
-    expense_id, _, date, cost, category_id, description = get_expense(
+    user = utils.get_user(user_id)
+    expense_id, _, date, cost, category_id, description = utils.get_expense(
         user[0], expense_id)
-    category = get_category(user[0], category_id)
+    category = utils.get_category(user[0], category_id)
     return {
         "cost": cost,
         "date": date,
@@ -96,7 +62,7 @@ class ExpenseJson(BaseModel):
 
 
 @router.post("/user/{user_id}/expense/", tags=["movies"])
-def add_conversation(user_id: int, expense: ExpenseJson):
+def add_expense(user_id: int, expense: ExpenseJson):
     """
     This endpoint adds a new expense to the database.
     This expense includes some required data and some optional data:
