+++ src/api/server.py
@@ -1,5 +1,5 @@
 from fastapi import FastAPI
-from src.api import budget, users
+from src.api import budget, users, expenses
 
 
 description = ""
@@ -19,6 +19,7 @@ app = FastAPI(
 )
 app.include_router(users.router)
 app.include_router(budget.router)
+app.include_router(expenses.router)
 
 @app.get("/")
 async def root():
